requirements.txt
pandas
numpy
tensorflow-gpu
gym
glob